feat: implement and apply Points Object
