+++ src/test/java/nextstep/ladder/domain/LineTest.java
@@ -9,9 +9,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LineTest {
 
     @Test
-    @DisplayName("플레이어의 수를 인자로 받아 사다리 라인을 생성한다.")
+    @DisplayName("좌표 값들을 인자로 받아 사다리 라인을 생성한다.")
     public void create() throws Exception {
-        Line line = new Line(new CountOfPlayer(4), new RandomConnectStrategy());
-        assertThat(line.points()).hasSize(4);
+        Line line = new Line(Points.createWith(new CountOfPlayer(4), new RandomConnectStrategy()));
+        assertThat(line.points().points()).hasSize(4);
     }
 }
